// src/types.ts
export interface Person {
    jmeno: string;
    prijmeni: string;
    email: string;
    plat: number;
    vztah: boolean;
    deti: number;
    datumNarozeni: string;
    vek: number;
    zkusenosti: string;
    vedlejsiPrijmy: number;
    konicky: string[];
    ulice: string;
    mesto: string;
    cisloPopisne: string;
    telefon: string;
    psc: string;
  }